+++ src/globals.py
@@ -1,2 +1,2 @@
-build=71
+build=72
 type='dev'